some changes without necessary results
